bpipe_dna_stages: Fix problem with forwarding input3.fastq in alignMEM
For haloplexHS I need to forward the third fastq file (index file)
which does not exist for other pipelines. Previously, I had
incorporated an if-clause to check whether forwarding was
necessary based on input.size()

However, this was not working reliably so I created distinct
alignment stages with and without that forward clause. It's
not pretty but it should work.