getCoveragePerGene.R: Fix ambiguous annotation, BED formats + coord padding
The script was failing for Anna because one specific gene in her
BED file was located on chrX and chrY.

I am now excluding all genes that are mapped to more than one chr

Anna's BED files also had 8 instead of 10 columns so another error
was thrown when labeling columns with a fixed string vector.

I fixed this, offering columnnames vectors for 10 and 8 column
BED files, testing for it with an if-clause.A

I also removed the +/- 5 bp padding of sequences in the BED file.
Montse's samples always had these additional bp covered in her
samples though they were not part of the covered design. Anna's
BED files matches her samples perfectly, so it seems padding is
not required.