Fix DESeq2* -> as.numeric(args)
Previously, args were assigned to variables as is. For LFC and
alpha cutoffs, values were taken as strings instead of numbers.

LFC would fail with a cryptic error when passed to DESeq2 so I
reassigned this within the script from the very beginning (and
never knew why...)

logFC did not fail but messed up comparisons! That is why the
number of DEGs reported in *summary.txt was always lower than the
number present in the *DESeq2results* files. Unfortunately, DEG
plots, both PCA and heatmaps, were based on the DEGs recognized by
summary(), so all of these plots are indeed incorrect and must be
recreated!

Example:
0.6  < "0.6"  -> F
0.5  < "0.6"  -> T
0.55 < "0.6"  -> T
-1   < "0.6"  -> F
-0.6 < "0.6"  -> T
1    < "0.6"  -> F

Fixing this bug also led to my file of the normalizedTransformedCounts
of DEGs finally being written to their own file as well. I added this
to the wrapper for annotation.

I also simplified the call to summary() to use the results-object
as is and without parameters. According to the DESeq2 manual, this
means the cutoff used for testing will also be used by the summary()
function.