Fix bpipe_finish: Test CSVs exist before trying to fix format
To collect data from manual inspections, I prepended columns to all
CSV files in the last commit. Unfortunately, I forgot to add a test
whether these CSV files exist at all - they do not for the standard
RNA pipeline which would cause the pipeline to fail without such a test.

I added an if clause to check for results_csv/ to prevent this