DESeq2: Add design to output prefix
Previously, the output file names were based only
on the input file prefix. Now, it's a combination
of the input file prefix and the full design
formula used. This should make results more
descriptive and prevent mixing up results
from different tests. Also, only with this
change can the same input file be analyzed with
different designs within the same folder without
overwriting previously written output files!

For clearer code, both scripts contain
an explicit variable for the output file prefix
instead of pasting this together based on the input
for each file written.

For additional consistency, the output file prefix
is created in the wrapper script and passed to the
R script. Previously, both scripts independently
pasted the prefix together, requiring changes to
be manually transmitted to both files each time
(and leaving output files with different names
when this was forgotten...)