DESeq2: Add & require param for fold change reference level
The third parameter passed to the DESeq2 wrapper script now
must be the reference level that all fold changes are to be
based on. Previously, the first level in alphabetical order
was selected. Results were seriously misinterpreted, when
this was not kept in mind...

Require both the design formula and the reference level to
be passed as input parameters by the user. Previously,
"~ condition" was the available default for the design.