DESeq2: Do not plot counts of too many factor combinations
The basic gene counts plot will be crowded and illegible.
Instead, calculate the number of factor level combinations
present in the design and plot only factors with up to
four combinations of levels
(such as treated / untreated * dx/rl).
For more complex designs, plot the final term of the design
formula only.