DESeq2: Add gene count plots for candidate genes
Read in a list of candidate genes' Ensemble IDs
for which counts are provided to DESeq2.
For each of these, print both the basic and pretty
plot of the gene's counts to a separate file.

Candidate genes are provided in "candidates.txt",
which must reside in the same directory as the
design table, and can not (yet) be passed as
a commandline argument.