Update get_flanking_sequence.sh
It turned out I was not very careful and the capitalization was affecting also other columns of the csv file. Now, I corrected this using process substitution on lines 150 and 152 like you did in the get_flanking_sequence_online.sh script, a very interesting feature that I was previously not aware of. However, to make this work, I had to change #!/bin/sh to #!/bin/bash. I hope this is not a problem. I tested the script with a big csv file and the results are now really identical with the output of the get_flanking_sequence_online.sh script (tested with diff).